# openAPI/openapi.oas3.yml
openapi: 3.0.3
info:
    title: III-Exhibition works-HAI
    description: API specification for interactions between Frontend, Backend, and Cloud Storage.
    version: 1.0.0
servers:
  - url: /api/handler
    description: Base path for all endpoints
paths:
    /api/handler/view:
        get:
            summary: Drawing Viewer Page
            description: Viewer Page for human AI drawings. 
            responses:
                "200":
                    description: Drawings were successfully fetched.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    result:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                position:
                                                    type: object
                                                    properties:
                                                        x:
                                                            type: integer
                                                            description: target of x position
                                                            example: 0
                                                        y:
                                                            type: integer
                                                            description: target of y position
                                                            example: 0
                                                    required:
                                                        - x
                                                        - y
                                                url:
                                                    type: string
                                                    description: "Presigned URL for target position."
                                            required:
                                                - position
                                                - url
                                required:
                                    - result

                            example:
                                result: [
                                    {
                                        position: {
                                            x: 0,
                                            y: 0
                                        },
                                        url: "https://storage.googleapis.com/..."
                                    },
                                    {
                                        position: {
                                            x: 0,
                                            y: -1
                                        },
                                        url: "https://storage.googleapis.com/..."
                                    }
                                
                                ]
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: "Invalid input data"
                "404":
                    description: Not Found - No drawing found for the specified position.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        description: "A detailed error message."
                                example:
                                    error: "No drawing found for the specified position."
                "500":
                    description: Internal Server Error - The server encountered an unexpected condition that prevented it from fulfilling the request.
                default:
                    description: Unexpected error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        description: "A detailed error message."

components:
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            in: header
            name: X-Api-Key
security:
    - ApiKeyAuth: []